+++ src/ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: kpastuhox <kirillpastuhow@gmail.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/07 14:35:30 by kpastuhox         #+#    #+#             */
-/*   Updated: 2019/09/05 13:08:12 by kpastuhox        ###   ########.fr       */
+/*   Updated: 2019/11/03 21:22:11 by kpastuhox        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-int		ft_strncmp(char const *s1, char const *s2, unsigned int n)
+int		ft_strncmp(char const *s1, char const *s2, size_t n)
 {
 	int i;
